scene/ui: fix contradictory comment, dedupe doc, drop cast noise
Final-pass cleanup (no behavior change):
- ui/text_panel.c: the fake-bold comment said "EXPERIMENT ... alpha
  0.25" but it's a shipped path and the code uses 0.15f. Dropped the
  stale EXPERIMENT tag and corrected the alpha in the comment.
- scene/backdrop.c: removed the second star-band-split comment that
  restated the STAR_BAND_CUT_* cutoffs documented 270 lines above
  (drift hazard); replaced with a pointer to the constants.
- ui/variable_panel.c, ui/profile_panel.c: collapsed redundant
  (float)((float)x) double casts to single casts (matches the
  prevailing idiom in the same files).

Verified-but-skipped (reclassified report-only, not clear-wins):
the color_picker.c checkerboard "duplicate" blocks differ in their
Y-origin math (top-anchored descending vs bottom-anchored ascending)
so a shared helper is a behavior-risk refactor, not a textual
factor-out; and the 0xAAAA line-stipple sites must NOT route through
SCENE_OCCLUDED_GHOST_STIPPLE (that constant is 0x0F0F — a different
dash pattern).

+++ src/ui/profile_panel.c
@@ -238,7 +238,7 @@ void ui_profile_panel_render(const UiRenderSnapshot *snap) {
     glEnable(GL_BLEND);
     glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
     ui_clr_a(UI_TOK_SUNKEN, 0.91f);
-    glRectf((float)((float)panel_x), (float)((float)panel_y), (float)((float)panel_x)+(float)((float)PROF_PANEL_W), (float)((float)panel_y)+(float)panel_h);
+    glRectf((float)panel_x, (float)panel_y, (float)panel_x + (float)PROF_PANEL_W, (float)panel_y + (float)panel_h);
 
     /* Border */
     ui_clr_a(UI_TOK_BORDER, 0.85f);
